% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMix.R
\name{DistMixing}
\alias{DistMixing}
\title{funkcja generujaca probe pochodzaca z 2 zadanych rozkladow}
\usage{
DistMixing(n, family_1, par_1, family_2, par_2, p = 0.5, ...)

DistMixing(n, family_1, par_1, family_2, par_2, p = 0.5, ...)
}
\arguments{
\item{n}{dlugosc proby}

\item{family_1}{nazwa pierwszego rozkladu}

\item{par_1}{parametry pierwszego rozkladu}

\item{family_2}{nazwa drugiego z rozkladow}

\item{par_2}{parametry drugiego z rozkladow}

\item{p}{czesc koncowej proby jaka maja stanowic obserwacje wylosowane z 1 rozkladu}

\item{...}{not sure}
}
\description{
funkcja generujaca probe pochodzaca z 2 zadanych rozkladow

funkcja generujaca probe pochodzaca z 2 zadanych rozkladow
}
