% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormTests.R
\name{NormalityTesting.default}
\alias{NormalityTesting.default}
\title{funkcja przeprowadzajaca testy normalnosci dla zadanej proby.}
\usage{
\method{NormalityTesting}{default}(x)
}
\arguments{
\item{x}{proba poddawana testom normalnosci}
}
\value{
Funkcja zwraca p.value dla 5 testow normalnosci:
\itemize{
    \item Shapiro
    \item JarqueBera
    \item AndersonDarling
    \item KolmogorovSmirnov
    \item PearsonChiSquare
}
}
\description{
funkcja przeprowadzajaca testy normalnosci dla zadanej proby.
}
