% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormTests.R
\name{NormalityTesting.DistMixing}
\alias{NormalityTesting.DistMixing}
\title{funkcja przeprowadzajaca testy normalnosci dla obiektu klasy DistMixing}
\usage{
\method{NormalityTesting}{DistMixing}(x)
}
\arguments{
\item{x}{obiekt klasy DistMixing}
}
\value{
Funkcja zwraca p.value dla 5 testow normalnosci:
\itemize{
    \item Shapiro
    \item JarqueBera
    \item AndersonDarling
    \item KolmogorovSmirnov
    \item PearsonChiSquare
}
}
\description{
funkcja przeprowadzajaca testy normalnosci dla obiektu klasy DistMixing
}
