% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormTests.R
\name{NormalityTesting}
\alias{NormalityTesting}
\title{funkcja przeprowadzajaca 5 testow normalnosci: Shapiro-Wilka, Jarque-Bera, Andersona-Darlinga, Kolmogorova-Smirnova oraz test Chi Pearsona dla zadanej proby.}
\usage{
NormalityTesting(x, ...)

NormalityTesting(x, ...)
}
\arguments{
\item{x}{proba poddawana testom normalnosci}

\item{...}{not sure}
}
\description{
funkcja przeprowadzajaca 5 testow normalnosci: Shapiro-Wilka, Jarque-Bera, Andersona-Darlinga, Kolmogorova-Smirnova oraz test Chi Pearsona dla zadanej proby.

funkcja generyczna
}
