% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMix.R
\name{print.DistMixing}
\alias{print.DistMixing}
\title{Komentarz}
\usage{
\method{print}{DistMixing}(x, ...)
}
\arguments{
\item{x}{obiekt klasy DistMixing, ktory chcemy wypisac}

\item{...}{not sure}
}
\description{
Komentarz
}
