% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMix.R
\name{summary.DistMixing}
\alias{summary.DistMixing}
\title{Metoda obliczajaca podstawowe statystyki opisowe}
\usage{
\method{summary}{DistMixing}(object, ...)
}
\arguments{
\item{object}{obiekt klasy DistMixing}

\item{...}{dalsze argumenty przekazywane do lub z innych metod}
}
\description{
Metoda obliczajaca podstawowe statystyki opisowe
}
