% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMix.R
\name{print.DistMixing}
\alias{print.DistMixing}
\title{Metoda wypisujaca obiekt klasy DistMixing}
\usage{
\method{print}{DistMixing}(x, ...)
}
\arguments{
\item{x}{obiekt klasy DistMixing, ktory chcemy wypisac}

\item{...}{dalsze argumenty przekazywane do lub z innych metod}
}
\description{
Metoda wypisujaca obiekt klasy DistMixing
}
