% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormTests.R
\name{print.NormalityTesting}
\alias{print.NormalityTesting}
\title{funkcja sluzaca do wypisywania wynikow NormalityTesting}
\usage{
\method{print}{NormalityTesting}(x, ...)
}
\arguments{
\item{x}{obiekt klasy NormalityTesting, ktory nalezy wypisac}

\item{...}{dalsze argumenty przekazywane do lub z innych metod}
}
\description{
funkcja sluzaca do wypisywania wynikow NormalityTesting
}
