% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMix.R
\name{summary.DistMixing}
\alias{summary.DistMixing}
\title{Komentarz}
\usage{
\method{summary}{DistMixing}(object, ...)
}
\arguments{
\item{object}{obiekt klasy DistMixing, dla ktorego chcemy obliczyc podst. statystyki opisowe}

\item{...}{not sure}
}
\description{
Komentarz
}
